feat: make it possible to configure the config/context file location
To make the rhoas-cli more friendly in regard of others cli that may be
created on top of it, it should be possible to set a custom location for
the configuration and context files. Before this commit, it was not
possible as the env var and the path where to store such files were
hardcoded which could elad to issues other clis based on the rhoc-cli
code would generate conflicts as they would act upon the same files.

This commit is about using adding support to set a sort of application
id that can be then used to determine the name of the env vars and to
compose the path for the config and context files.