chore: remove useless checks
The Factory's connection function is in charge of configuring how to
connect to the service and to configure the related builder a number of
checks were perfomed against empty strings.

Given that an empty string is also the default string value, there's no
value to perform such check as the conenction builder fields if type
string are also initialized to an empty string by default.